Fixed data type in setStreamVersion

+++ src/info/ata4/unity/assetbundle/AssetBundleHeader.java
@@ -150,7 +150,7 @@ public class AssetBundleHeader implements Struct {
         return streamVersion;
     }
 
-    public void setStreamVersion(byte format) {
+    public void setStreamVersion(int format) {
         this.streamVersion = format;
     }
